% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kin_dat.R
\name{calculate_replicate_state_uptake}
\alias{calculate_replicate_state_uptake}
\title{Calculcate uptake}
\usage{
calculate_replicate_state_uptake(
  rep_peptide_dat,
  state = dat[["State"]][1],
  time_0 = min(dat[["Exposure"]]),
  time_100 = max(dat[["Exposure"]])
)
}
\arguments{
\item{rep_peptide_dat}{data from experimental file, but aggregated within
technical replicates}

\item{state}{biological state}

\item{time_0}{minimal exchange control time point}

\item{time_100}{maximal exchange control time point}
}
\value{
a data.frame object.
}
\description{
Function calculates deuterim uptake for one biological state,
for replicates, without additional aggregation of the data.
}
\examples{
dat <- HaDeX::read_hdx(...)
TODO

}
\seealso{
HaDeX::read_hdx
}
