% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_uc.R
\name{calculate_uc_from_hires_peptide}
\alias{calculate_uc_from_hires_peptide}
\title{Recovers deuterium uptake from hi-res parameters}
\usage{
calculate_uc_from_hires_peptide(
  fit_dat,
  fit_values_all,
  hires_dat = NULL,
  fractional = TRUE,
  hires_method = c("shortest", "weighted")
)
}
\arguments{
\item{fit_dat}{uptake data filtered for given peptide}

\item{fit_values_all}{fit parameters for whole peptide pool, unaggregated}

\item{fractional}{indicator if fractional values are used}

\item{hires_method}{method of aggregation}
}
\description{
This function recovers fractional deuterium uptake from high-resolution level
back to the peptide level. First, the values are aggregated into high-resolution
level using selected aggregation method.
If there are no calculated recovered values, probably the peptide was calssified as
invalid - check that in fit_values_all param to be sure.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
fit_values_all <- create_fit_dataset(kin_dat, control = get_example_control(),
                                     fit_k_params = get_example_fit_k_params_2(),
                                     fractional = TRUE)
fit_dat <- kin_dat[kin_dat[["ID"]] == 54, ]
calculate_uc_from_hires_peptide(fit_dat, fit_values_all, hires_method = "weighted")

}
\seealso{
\code{\link{create_uc_from_hires_dataset}}
}
