% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_uc.R
\name{compare_aggregation_methods}
\alias{compare_aggregation_methods}
\title{Compare aggregation methods}
\usage{
compare_aggregation_methods(
  kin_dat,
  fit_values_all,
  style = c("coverage", "butterfly")
)
}
\arguments{
\item{kin_dat}{uptake data for the whole peptide pool}

\item{fit_values_all}{fit parameters for whole peptide pool, unaggregated}

\item{style}{plotting style}
}
\description{
The function plots the analogical plots as plot_recovered_uc_coverage but
for the two aggregation methods at the same time.
TODO: rewrite for optimization purposes.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat, state = "Alpha_KSCN")
fit_values_all <- create_fit_dataset(kin_dat, control = get_example_control(),
                                     fit_k_params = get_example_fit_k_params(),
                                     fractional = T)
compare_aggregation_methods(kin_dat, fit_values_all, style = "coverage")
compare_aggregation_methods(kin_dat, fit_values_all, style = "butterfly")
}
