% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fit_dataset.R
\name{create_fit_dataset}
\alias{create_fit_dataset}
\title{Create fits for the whole peptide pool}
\usage{
create_fit_dataset(
  kin_dat,
  fit_k_params,
  control = list(maxiter = 1000, scale = "levenberg"),
  trace = FALSE,
  fractional = FALSE,
  omit_t_100 = FALSE,
  workflow = 321
)
}
\arguments{
\item{kin_dat}{calculated uptake data for
one biological state, result of e.q.
\code{\link{prepare_kin_dat}}}

\item{fit_k_params}{boundaries for exchange
groups, example in prepare_kin_dat}

\item{control}{control options for the fitting process,
example in \code{\link{get_example_control}}}

\item{trace}{logical, indicator if fitting trace is to
be displayed}

\item{fractional}{logical, indicator if normalized values
are used in the fitting process}

\item{workflow}{workflow type, options: 321, 31, 21, indicating
the types of functions used in fitting process}
}
\value{
a data.frame object.
}
\description{
This function calls the fitting function \code{\link{get_fit_results}}
for each peptide in supplied kin_dat creating the dataset for whole
peptide pool.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
fit_k_params <- get_example_fit_k_params()
head(create_fit_dataset(kin_dat, fit_k_params))

}
\seealso{
\code{\link{get_fit_results}}
\code{\link{prepare_kin_dat}}
\code{\link{get_example_fit_k_params}}
\code{\link{get_example_control}}
}
