% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kin_dat.R
\name{create_replicate_state_uptake_dataset}
\alias{create_replicate_state_uptake_dataset}
\title{Creates deuterium uptake replicate dataset}
\usage{
create_replicate_state_uptake_dataset(
  rep_dat,
  state = dat[["State"]][1],
  time_0 = min(dat[["Exposure"]]),
  time_100 = max(dat[["Exposure"]])
)
}
\arguments{
\item{rep_dat}{data}

\item{state}{biological state}

\item{time_0}{minimal exchange control time point}

\item{time_100}{maximal exchange control time point}
}
\value{
a data.frame object.
}
\description{
Wrapper for calculate_replicate_state_uptake, interating
throught all the peptides in provided data.
}
\examples{
TODO

}
\seealso{
calculate_replicate_state_uptake
}
