% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_uc.R
\name{create_uc_from_hires_dataset}
\alias{create_uc_from_hires_dataset}
\title{Creates recovered deuterium uptake dataset}
\usage{
create_uc_from_hires_dataset(
  kin_dat,
  fit_values_all,
  fractional = T,
  hires_method = c("shortest", "weighted")
)
}
\arguments{
\item{kin_dat}{uptake data for the whole peptide pool}

\item{fit_values_all}{fit parameters for whole peptide pool, unaggregated}

\item{fractional}{indicator if fractional values are used}

\item{hires_method}{method of aggregation}
}
\description{
Wrapper for \code{\link{calculate_uc_from_hires_peptide}} to create the data
for whole pepitde pool.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat, state = "Alpha_KSCN")
fit_values_all <- create_fit_dataset(kin_dat, control = get_example_control(),
                                     fit_k_params = get_example_fit_k_params(),
                                     fractional = T)
create_uc_from_hires_dataset(kin_dat, fit_values_all, hires_method = "shortest")

}
\seealso{
\code{\link{calculate_uc_from_hires_peptide}}
}
