% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_chimera_commands.R
\name{export_chimera_commands}
\alias{export_chimera_commands}
\title{Export coloring commands for ChimeraX}
\usage{
export_chimera_commands(
  hires_params,
  dominant = FALSE,
  state = hires_params[["State"]][1],
  chain = "A"
)
}
\description{
Export coloring commands for ChimeraX
}
