% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fit_results.R
\name{get_fit_results}
\alias{get_fit_results}
\title{Uptake curve fitting function}
\usage{
get_fit_results(
  fit_dat,
  fit_k_params,
  control = list(maxiter = 1000, scale = "levenberg"),
  trace = FALSE,
  workflow = 31,
  fractional = TRUE,
  omit_t_100 = FALSE,
  edge_times = c(min(fit_dat[["Exposure"]]), max(fit_dat[["Exposure"]]))
)
}
\arguments{
\item{fit_dat}{uptake data used for fit,
filtered uptake data for specific peptide}

\item{fit_k_params}{boundaries for exchange
groups, example in prepare_kin_dat}

\item{control}{control options for the fitting process,
example in \code{\link{get_example_control}}}

\item{trace}{logical, indicator if fitting trace is to
be displayed}

\item{workflow}{workflow type, options: 321, 31, 21, indicating
the types of functions used in fitting process}

\item{fractional}{logical, indicator if normalized values
are used in the fitting process}

\item{omit_t_100}{logical, indicator if the measurement associated
with t_100 is included in the fitting process or not}

\item{edge_times}{edge measurement times of the uptake curve}
}
\value{
a data.frame object.
}
\description{
This function analyses the supplied deuterium uptake curve
for specific peptide and provides the fitting results or
assigned class.
Firstly, based on the absolute uptake values, the function
check if the data fulfills requirements to be assigned
a class (invalid, none exchange). If not, it is possible
to fit a function in a specific workflow.
To create results for more than one peptide, see
\code{\link{create_fit_dataset}}
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
fit_dat <- kin_dat[kin_dat[["ID"]]==1,  ]
fit_k_params <- get_example_fit_k_params()
get_fit_results(fit_dat, fit_k_params)

}
