% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peptide_list}
\alias{get_peptide_list}
\title{Provides a peptide list}
\usage{
get_peptide_list(dat)
}
\arguments{
\item{dat}{data.frame}
}
\value{
a data.frame with position values, Start and End.
}
\description{
Created peptite list based on the provided data.frame.
}
\examples{
dat <- HaDeX2::read_hdx(...)
get_peptide_list(dat)

}
