% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_peptide_list_v2}
\alias{get_peptide_list_v2}
\title{Provides a peptide list}
\usage{
get_peptide_list_v2(dat_1, dat_2)
}
\arguments{
\item{dat_1}{data.frame}

\item{dat_2}{data.frame}
}
\value{
a data.frame with sequence and position values.
}
\description{
Created peptite list from two binded data.frames.
The peptides are selected based on their position values. If there
are two peptides with the same position but different sequence, both sequences
are shown in the final data.frame.
}
\examples{
dat_1 <- HaDeX2::read_hdx(...)
dat_2 <- HaDeX2::read_hdx(...)
get_peptide_list_v2(dat_1, dat_2)

}
