% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getters.R
\name{get_residue_positions}
\alias{get_residue_positions}
\title{Get a data table with positions and residues}
\usage{
get_residue_positions(fit_values)
}
\description{
Get a data table with positions and residues
}
