% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hradexify.R
\name{hradexify}
\alias{hradexify}
\title{HRaDeX customized ggplot theme}
\usage{
hradexify(plt)
}
\arguments{
\item{plt}{ggplot object. Plot to HRaDeXify.}
}
\value{
a \code{\link{ggplot2}} object.
}
\description{
This function HRaDeXifies plot. It adds HRaDeX logo
and ggplot theme.
}
\details{
Function adds the logo of HRaDeX package in the left down
corner of the plot and the HRaDeX theme.
This function is adapted from the HaDeX package.
}
