% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit_amino.R
\name{omit_amino}
\alias{omit_amino}
\title{Omit first amino of peptides}
\usage{
omit_amino(dat, omit = 1)
}
\arguments{
\item{dat}{dataset to be modified}

\item{omit}{numer of residues to me omitted for each peptide}
}
\value{
data.frame in the same format as provided.
}
\description{
In some approach, ommitting the first amino residue in the peptide is a
way to correct back-exchange. This function modifies provided dataset and omits desired
number of residues (omit parameter) effectively changing the Start position of each
peptide.
}
\examples{
fixed_dat <- omit_amino(alpha_dat)
head(fixed_dat)

}
