% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_structure_diff.R
\name{plot_3d_structure_blank}
\alias{plot_3d_structure_blank}
\title{Plots blank 3D structure}
\usage{
plot_3d_structure_blank(pdb_file_path)
}
\arguments{
\item{pdb_file_path}{...}
}
\value{
a r3dmol object.
}
\description{
Plots blank 3D strucutre of the protein. This function is the base line
for the presentation of the differences in compahradex.
}
\seealso{
plot_3d_structure_hires
}
