% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_structure_dominant.R
\name{plot_3d_structure_dominant}
\alias{plot_3d_structure_dominant}
\title{Plot dominant class on 3D structure}
\usage{
plot_3d_structure_dominant(hires_params, pdb_file_path)
}
\description{
Plot dominant class on 3D structure
}
\examples{
pdb_file_path <- paste0(system.file(package = "HRaDeX"), "/data/Model_eEF1Balpha.pdb")
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
kin_dat <- prepare_kin_dat(alpha_dat)
fit_values <- create_fit_dataset(kin_dat = kin_dat, control = control, fit_k_params = fit_k_params, fractional = TRUE)
hires_params <- calculate_hires(fit_values, method = "weighted", fractional = TRUE)
plot_3d_structure_dominant(hires_params = hires_params, pdb_file_path = pdb_file_path)

}
