% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_structure_hires.R
\name{plot_3d_structure_hires}
\alias{plot_3d_structure_hires}
\title{Presents classification results on 3D structure}
\usage{
plot_3d_structure_hires(hires_params, pdb_file_path)
}
\arguments{
\item{hires_params}{...}

\item{pdb_file_path}{...}
}
\value{
a r3dmol object.
}
\description{
Plots the color classification on supplied 3d structure.
}
\examples{
dat <- HaDeX::read_hdx(...)
pdb_file_path <- ...
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
kin_dat <- prepare_kin_dat(dat)
fit_values <- create_fit_dataset(kin_dat, control, fit_k_params)
hires_params <- calculate_hires(fit_values)
plot_3d_structure_hires(hires_params, pdb_file_path)

}
\seealso{
calculate_hires
}
