% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_class_heatmap}
\alias{plot_class_heatmap}
\title{Plots heatmap with classification parameters}
\usage{
plot_class_heatmap(fixed_params)
}
\arguments{
\item{fixed_params}{fit params after modification}
}
\value{
a ggplot2 object.
}
\description{
Not in use.
TODO
}
