% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cov_class.R
\name{plot_cov_class}
\alias{plot_cov_class}
\title{Plots coverage with classification results}
\usage{
plot_cov_class(fit_values, fractional = T, interactive = F)
}
\arguments{
\item{fit_values}{...}

\item{fractional}{...}

\item{interactive}{...}
}
\value{
a girafe object.
}
\description{
This function is used for visualization of the classification method
for peptides. It uses the traditional form of coverage, with length present and easily
estimated redundancy. It also allows to see if the tendencies for exchange are coherent
in regions.
}
\examples{
HaDeX::read_hdx(...)
kin_dat <- prepare_kin_dat(dat, state = state_1)
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
fit_values <- create_fit_dataset(kin_dat, control, fit_k_params)
plot_cov_class(fit_values)

}
\seealso{
create_fit_dataset
}
