% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_double_uc.R
\name{plot_double_uc}
\alias{plot_double_uc}
\title{Plots two uc plots side by side}
\usage{
plot_double_uc(fit_dat, fit_values, replicate = F, fractional = T)
}
\arguments{
\item{fit_dat}{uptake data for selected peptide.}

\item{fit_values}{fit values for selected peptide.}

\item{replicate}{...}

\item{fractional}{...}
}
\value{
a ggplot object.
}
\description{
This function plots two plots side by side for a peptide. The left
plot is uc plot with fitted model with its components. The right plot is singular
uc without logarithmic axis for the comparison of shape.
Basically, this function calls for other functions and plots the results side by side.
This function is aimed for the summary \code{Plots} panel in HRaDeXGUI and does not have
an interactive mode.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
fit_dat <- kin_dat[kin_dat[["ID"]]==1, ]
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
fit_values_all <- create_fit_dataset(kin_dat, fit_k_params, control)
fit_values <- fit_values_all[fit_values_all[["ID"]]==1, ]
plot_double_uc(fit_dat, fit_values)

}
\seealso{
plot_fitted_uc plot_singular_uc
}
