% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_estimated_k.R
\name{plot_estimated_k}
\alias{plot_estimated_k}
\title{Plots estimated k}
\usage{
plot_estimated_k(hires_params, interactive = F)
}
\arguments{
\item{hires_params}{data.frame with hires results, calculate using
calculate_hires.}

\item{interactive}{...}
}
\value{
a girafe object.
}
\description{
Plots the estimated k for residues. Estimated k for each are calculated
from classification results, treating the n (population) as the probability of getting
corresponding rate of exchange. The aggregation of results is described in the vignette.
}
\examples{
dat <- HaDeX::read_hdx(...)
kin_dat <- prepare_kin_dat(dat)
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
fit_values <- create_fit_dataset(kin_dat, control, fit_k_params)
hires_params <- calculate_hires(fit_values)
plot_estimated_k(hires_params)

}
\seealso{
calculate_hires
}
