% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fitted_uc.R
\name{plot_fitted_uc}
\alias{plot_fitted_uc}
\title{Plots uptake curve with fit components}
\usage{
plot_fitted_uc(
  fit_dat,
  fit_values,
  replicate = F,
  fractional = T,
  interactive = F
)
}
\arguments{
\item{fit_dat}{uptake data for selected peptide.}

\item{fit_values}{fit values for selected peptide.}

\item{replicate}{indicator if the replicate data is plotted, or aggregated.}

\item{fractional}{...}

\item{interactive}{...}
}
\value{
a ggplot object. Need converstion to girafe for interactivity.
}
\description{
FUnction plots the uptake values for one, selected peptide. Alongside the
fitted model is plotted in black, with its components (fast - red, medium - green,
slow - blue). In case when peptide is classified as a edge case, the linear model is used.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
fit_dat <- kin_dat[kin_dat[["ID"]]==1, ]
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
fit_values_all <- create_fit_dataset(kin_dat, control, fit_k_params)
fit_values <- fit_values_all[fit_values_all[["ID"]]==1, ]
plot_fitted_uc(fit_dat, fit_values)

}
\seealso{
plot_lm
}
