% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hires.R
\name{plot_hires}
\alias{plot_hires}
\title{Plots high resolution classification plot}
\usage{
plot_hires(hires_params, interactive = F)
}
\description{
Plots high resolution classification plot
}
\examples{
dat <- HaDeX::read_hdx()

}
