% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_states_distance.R
\name{plot_k_distance}
\alias{plot_k_distance}
\title{Plots estimated k distance}
\usage{
plot_k_distance(two_state_dataset, interactive = FALSE)
}
\arguments{
\item{two_state_dataset}{data.frame produced by \link{create_two_state_dataset}}

\item{interactive}{logical, true for GUI}
}
\value{
a ggiraph object.
}
\description{
Function plots the difference between two estimated k values from
hires parameters for two biological states. The sign of difference indicates the
direction of change.
}
\examples{
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
kin_dat <- prepare_kin_dat(alpha_dat)
fit_values <- create_fit_dataset(kin_dat, control, fit_k_params)
hires_params <- calculate_hires(fit_values)
# the same for the second state and then:
# two_states_dataset <- create_two_state_dataset(hires_params_1, hires_params_2)
# plot_k_distance(two_states_dataset)

}
\seealso{
create_two_state_dataset
}
