% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_lm.R
\name{plot_lm}
\alias{plot_lm}
\title{Plots uc with linear fit}
\usage{
plot_lm(fit_dat, class_name = NA, interactive = T)
}
\arguments{
\item{fit_dat}{deuterium uptake data for selected peptide.}

\item{class_name}{class name assigned to selected peptide.}

\item{interactive}{...}
}
\value{
a ggplot2 object. For interactive mode, conversion to girafe is needed.
}
\description{
Plots uptake curve for a peptide with linear fit.
Used in GUI to present uptake curves for edge cases, where the
three-exponential model cannot be fitted.
}
\examples{
dat <- HaDeX::read_hdx()
kin_dat <- prepare_kin_dat(dat, state = state_1)
fit_dat <- kin_dat[kin_dat[["id"]]==1, ]
plot_lm(fit_dat, class_name = unique(fit_dat[["class_name"]]))

}
\seealso{

}
