% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_n.R
\name{plot_n}
\alias{plot_n}
\title{Plots n values}
\usage{
plot_n(list_params, fractional = FALSE, interactive = FALSE)
}
\arguments{
\item{list_params}{list of fit parameters}

\item{fractional}{...}

\item{interactive}{...}
}
\value{
a ggplot2 object. Need of convertion to girafe for interactive mode.
}
\description{
Plots the n values from fit parameters. When operating on fractional
values, the n value should be close to 1 to ensure model legitimacy. When operating
on values in daltons, the upper limit for the n value i max possible uptake calculated
from the sequence, although rarely reached due to back-exchange.
}
\examples{
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
kin_dat <- prepare_kin_dat(alpha_dat)
fit_values <- create_fit_dataset(kin_dat, control, fit_k_params)
plot_n(fit_values)

}
\seealso{
create_fit_results
}
