% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_params_map.R
\name{plot_params_map}
\alias{plot_params_map}
\title{Plots map of params}
\usage{
plot_params_map(fit_values, interactive = F)
}
\arguments{
\item{fit_values}{...}

\item{interactive}{...}
}
\description{
Currently not in use.
}
