% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rss_hist.R
\name{plot_rss_hist}
\alias{plot_rss_hist}
\title{Histogram of fit rss values}
\usage{
plot_rss_hist(fit_values, interactive = F)
}
\arguments{
\item{fit_values}{data.frame with fit values}

\item{interactive}{...}
}
\value{
a ggplot2 object. Needs converstion to girafe for the interactive mode.
}
\description{
Histogram of rss values for selected models for each peptide.
Divided with respect to the number of model parameters. Plot to judge the goodness
of the models.
}
\examples{
TODO

}
\seealso{
...
}
