% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_singular_uc.R
\name{plot_singular_uc}
\alias{plot_singular_uc}
\title{Plots singular uptake curve}
\usage{
plot_singular_uc(
  fit_dat,
  fit_values,
  include_uc = T,
  replicate = F,
  fractional = T,
  interactive = F
)
}
\arguments{
\item{fit_dat}{deuterium uptake data for one peptide}

\item{fit_values}{fit parameters for that one peptide, not used}

\item{include_uc}{indicator if regular uc should be plotted as well}

\item{replicate}{indicator if data for replicates should be plotted, or
aggregated}

\item{fractional}{...}

\item{interactive}{...}
}
\value{
a ggplot object. Need a conversion to girafe for interactive mode.
}
\description{
Function plots deuterium uptake plot (uc) for one peptide,
in a traditional mode.
Currently not in use in GUIs.
}
\examples{
TODO

}
\seealso{
...
}
