% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_states_distance.R
\name{plot_two_states}
\alias{plot_two_states}
\title{Plots the classification results for two states}
\usage{
plot_two_states(
  hires_params_1,
  hires_params_2,
  type = c("aggregated", "classes", "coverage"),
  interactive = F
)
}
\arguments{
\item{hires_params_1}{hires parameters for the first state}

\item{hires_params_2}{hires parameters for the second state}

\item{type}{not supported right now}

\item{interactive}{logical, true for GUI}
}
\value{
a ggiraph object.
}
\description{
Function plots the classfication results in color code for two biological
states at the same time. The results for the second state are on top, and for the
first state on the bottom. The grey color indicates no data.
}
\examples{
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
control <- get_example_control()
kin_dat <- prepare_kin_dat(alpha_dat)
fit_values <- create_fit_dataset(kin_dat, control, fit_k_params)
hires_params <- calculate_hires(fit_values)
# the same for the second state, and then:
# plot_two_states(hires_params_1, hires_params_2)

}
\seealso{
calculate_hires create_fit_dataset
}
