% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_states_distance.R
\name{plot_uc_distance}
\alias{plot_uc_distance}
\title{Plots the uc difference type 2}
\usage{
plot_uc_distance(
  uc_distance_dataset,
  squared = F,
  fractional = T,
  interactive = F
)
}
\arguments{
\item{uc_distance_dataset}{...}

\item{squared}{indicator if the uc distance is squared}

\item{fractional}{...}

\item{interactive}{logical, true for GUI}
}
\value{
a ggplot2 object.
}
\description{
Function plots the uc distance from uc_distance_dataset.
Be careful, as uc_distance_dataset contains distances calculated using
different methods, as described in vignette TODO.
}
\examples{
kin_dat_1 <- prepare_kin_dat(alpha_dat, state = unique(alpha_dat[["State"]])[1])
kin_dat_2 <- prepare_kin_dat(alpha_dat, state = unique(alpha_dat[["State"]])[2])
uc_distance_dataset <- create_uc_distance_dataset(kin_dat_1, kin_dat_2)
plot_uc_distance(uc_distance_dataset)

}
\seealso{
create_uc_distance_dataset
}
