% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_chimera_commands.R
\name{prepare_chimera_commands}
\alias{prepare_chimera_commands}
\title{Prepare coloring commands for ChimeraX}
\usage{
prepare_chimera_commands(
  hires_params,
  dominant = FALSE,
  state = hires_params[["State"]][1],
  chains = c("A")
)
}
\arguments{
\item{state}{biological state}

\item{chain}{chain for which the coloring
commands are created}
}
\description{
This function prepares the coloring commands for ChimeraX.
If you wish only to create coloring commands and downlaod them as
a text file, see \link{\code{export_chimera_commands}} function.
}
\examples{
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
kin_dat <- prepare_kin_dat(alpha_dat)
fit_values <- create_fit_dataset(kin_dat = kin_dat, control = control, fit_k_params = fit_k_params, fractional = TRUE)
hires_params <- calculate_hires(fit_values, method = "weighted", fractional = TRUE)

prepare_chimera_commands(hires_params, chain = "G")
prepare_chimera_commands(hires_params, chain = "G", dominant = TRUE)

}
