% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3d_structure_diff.R
\name{prepare_diff_data}
\alias{prepare_diff_data}
\title{Prepares differential data}
\usage{
prepare_diff_data(params, value, threshold)
}
\arguments{
\item{params}{values to be presented.}

\item{value}{which value should be presented.}

\item{threshold}{threshold for the values.}
}
\value{
a data.frame object.
}
\description{
Function prepares the results in a format that can be used
for 3D presentation. It transforms the peptide data into resiudal data, with regard
to the original values above the threshold.
}
\seealso{
plot_3d_structure_blank
}
