% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kin_dat.R
\name{prepare_kin_dat}
\alias{prepare_kin_dat}
\title{Creates uptake data from mass measurements}
\usage{
prepare_kin_dat(
  dat,
  state = dat[["State"]][1],
  time_0 = min(dat[["Exposure"]]),
  time_100 = max(dat[["Exposure"]]),
  replicate = FALSE,
  FD = FALSE
)
}
\arguments{
\item{dat}{experimental data read by HaDeX::read_hdx}

\item{state}{biological state}

\item{time_0}{minimal exchange control time point}

\item{time_100}{maximal exchange control time point}

\item{replicate}{idicator, if the replicate values should be provided, or aggregated}

\item{FD}{idicator, if the time_100 value shoul be used as fully deuterated control,
only for normalization purposes, or left as time point of measurement}
}
\value{
a data.frame object.
}
\description{
Function calculates deuterium uptake for each time point
in the original data with respect to the parameters.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
head(kin_dat)

}
\seealso{
HaDeX::read_hdx
}
