% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_uc.R
\name{recreate_uc}
\alias{recreate_uc}
\title{Plots recovered DU of a peptide}
\usage{
recreate_uc(
  fit_dat,
  fit_values_all,
  fractional = TRUE,
  hires_method = c("shortest", "weighted")
)
}
\arguments{
\item{fit_dat}{experimental uptake data filter for chosen peptide}

\item{fit_values_all}{fit parameters for whole peptide pool}

\item{fractional}{indicator if the fractional deuterium uptake is used}

\item{hires_method}{method of aggregation}
}
\description{
The function recovers deuterium uptake from high-resolution
parameters and returns a plot of uptake curve for chosen peptide.
TODO: rewrite as only plotting function
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat, state = "Alpha_KSCN")
fit_dat <- kin_dat[kin_dat[["ID"]]==3, ]
fit_values_all <- create_fit_dataset(kin_dat, control = get_example_control(),
                                     fit_k_params = get_example_fit_k_params(),
                                     fractional = T)
recreate_uc(fit_dat, fit_values_all, hires_method = "shortest")

}
