% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_uc.R
\name{calculate_uc_from_hires_peptide}
\alias{calculate_uc_from_hires_peptide}
\title{Recovers deuterium uptake from hi-res parameters}
\usage{
calculate_uc_from_hires_peptide(
  fit_dat,
  fit_values_all,
  fractional = T,
  hires_method = c("shortest", "weighted")
)
}
\arguments{
\item{fit_dat}{uptake data filtered for given peptide}

\item{fit_values_all}{fit parameters for whole peptide pool, unaggregated}

\item{fractional}{indicator if fractional values are used}

\item{hires_method}{method of aggregation}
}
\description{
This function recovers fractional deuterium uptake from high-resolution level
back to the peptide level. First, the values are aggregated into high-resolution
level using selected aggregation method.
}
\seealso{
\code{\link{create_uc_from_hires_dataset}}
}
