% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fit_dataset.R
\name{create_fit_dataset}
\alias{create_fit_dataset}
\title{Create fits for all the peptides}
\usage{
create_fit_dataset(
  kin_dat,
  fit_k_params,
  control = list(maxiter = 1000, scale = "levenberg"),
  trace = FALSE,
  fractional = FALSE,
  workflow = 321
)
}
\arguments{
\item{kin_dat}{...}

\item{fit_k_params}{...}

\item{control}{...}

\item{trace}{...}

\item{fractional}{...}

\item{workflow}{...}
}
\value{
a data.frame object.
}
\description{
This is a wrapper function for get_fit_results, but for multiple peptides at
the same time. First, it creates a peptide list from the supplied deuterium uptake data, and
calls the fitting function for each, to create full dataset.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
fit_k_params <- get_example_fit_k_params()
head(create_fit_dataset(kin_dat, fit_k_params))

}
\seealso{
get_fit_results
}
