% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kin_dat.R
\name{create_replicate_state_uptake_dataset}
\alias{create_replicate_state_uptake_dataset}
\title{TODO}
\usage{
create_replicate_state_uptake_dataset(
  rep_dat,
  state = dat[["State"]][1],
  time_0 = min(dat[["Exposure"]]),
  time_100 = max(dat[["Exposure"]])
)
}
\description{
TODO
}
