% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_states.R
\name{create_uc_distance_dataset}
\alias{create_uc_distance_dataset}
\title{Creates dataset with uc distances}
\usage{
create_uc_distance_dataset(kin_dat_1, kin_dat_2)
}
\arguments{
\item{kin_dat_1}{kinetic data for the first state}

\item{kin_dat_2}{kinetic data for the second state}
}
\value{
data.frame for plotting functions
}
\description{
This function iterates through common list of peptides in
two states, based on their position in the sequence. For each pair of peptide
data, the uc distance is calculated. The method of uc distance calucation is
described in the vigniette TODO.
}
\examples{
kin_dat_1 <- prepare_kin_dat(alpha_dat, state = unique(alpha_dat[["State"]])[1])
kin_dat_2 <- prepare_kin_dat(alpha_dat, state = unique(alpha_dat[["State"]])[2])
uc_dist_dataset <- create_uc_distance_dataset(kin_dat_1, kin_dat_2)
head(uc_dist_dataset)

}
\seealso{
get_uc_distance plot_uc_distance
}
