% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hires.R
\name{plot_hires}
\alias{plot_hires}
\title{Plots high resolution classification plot}
\usage{
plot_hires(hires_params, interactive = F)
}
\arguments{
\item{hires_params}{data.frame with hires results, calculate using
calculate_hires.}

\item{interactive}{...}
}
\value{
a girafe object.
}
\description{
Plot the result of color classification analysis in
high resolution in a linear form.
}
\examples{
kin_dat <- prepare_kin_dat(alpha_dat)
fit_k_params <- get_example_fit_k_params()
control <- get_example_control()
fit_values <- create_fit_dataset(kin_dat, control, fit_k_params)
hires_params <- calculate_hires(fit_values)
plot_hires(hires_params)

}
\seealso{
calculate_hires
}
